# contracts/openapi/ai-gateway-v1.json
{
  "components": {
    "schemas": {
      "AudioTranscriptionResponse": {
        "additionalProperties": false,
        "properties": {
          "detected_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Language"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageMeasurement"
          }
        },
        "required": [
          "request_id",
          "text",
          "provider",
          "model"
        ],
        "title": "AudioTranscriptionResponse",
        "type": "object"
      },
      "Body_analyze_media_ai_v1_media_analyses_post": {
        "properties": {
          "media": {
            "format": "binary",
            "title": "Media",
            "type": "string"
          },
          "product": {
            "pattern": "^[a-z][a-z0-9_-]{1,63}$",
            "title": "Product",
            "type": "string"
          },
          "profile": {
            "pattern": "^[a-z][a-z0-9_.-]{2,127}$",
            "title": "Profile",
            "type": "string"
          },
          "prompt": {
            "maxLength": 50000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "system_instruction": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "System Instruction",
            "type": "string"
          },
          "temperature": {
            "default": 0.2,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "type": "number"
          }
        },
        "required": [
          "request_id",
          "product",
          "profile",
          "system_instruction",
          "prompt",
          "media"
        ],
        "title": "Body_analyze_media_ai_v1_media_analyses_post",
        "type": "object"
      },
      "Body_transcribe_audio_ai_v1_audio_transcriptions_post": {
        "properties": {
          "audio": {
            "format": "binary",
            "title": "Audio",
            "type": "string"
          },
          "product": {
            "pattern": "^[a-z][a-z0-9_-]{1,63}$",
            "title": "Product",
            "type": "string"
          },
          "profile": {
            "pattern": "^[a-z][a-z0-9_.-]{2,127}$",
            "title": "Profile",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "source_language": {
            "default": "auto",
            "maxLength": 35,
            "minLength": 2,
            "title": "Source Language",
            "type": "string"
          },
          "terminology_prompt": {
            "default": "",
            "maxLength": 2000,
            "title": "Terminology Prompt",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "product",
          "profile",
          "audio"
        ],
        "title": "Body_transcribe_audio_ai_v1_audio_transcriptions_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MediaAnalysisResponse": {
        "additionalProperties": false,
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output": {
            "title": "Output",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageMeasurement"
          }
        },
        "required": [
          "request_id",
          "output",
          "provider",
          "model"
        ],
        "title": "MediaAnalysisResponse",
        "type": "object"
      },
      "SpeechSynthesisRequest": {
        "additionalProperties": false,
        "properties": {
          "input": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Input",
            "type": "string"
          },
          "instructions": {
            "default": "",
            "maxLength": 1000,
            "title": "Instructions",
            "type": "string"
          },
          "product": {
            "pattern": "^[a-z][a-z0-9_-]{1,63}$",
            "title": "Product",
            "type": "string"
          },
          "profile": {
            "pattern": "^[a-z][a-z0-9_.-]{2,127}$",
            "title": "Profile",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "voice": {
            "pattern": "^[A-Za-z0-9_-]{1,64}$",
            "title": "Voice",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "product",
          "profile",
          "input",
          "voice"
        ],
        "title": "SpeechSynthesisRequest",
        "type": "object"
      },
      "TextGenerationRequest": {
        "additionalProperties": false,
        "properties": {
          "input": {
            "maxLength": 200000,
            "minLength": 1,
            "title": "Input",
            "type": "string"
          },
          "product": {
            "pattern": "^[a-z][a-z0-9_-]{1,63}$",
            "title": "Product",
            "type": "string"
          },
          "profile": {
            "pattern": "^[a-z][a-z0-9_.-]{2,127}$",
            "title": "Profile",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "response_format": {
            "default": "text",
            "enum": [
              "text",
              "json"
            ],
            "title": "Response Format",
            "type": "string"
          },
          "system_instruction": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "System Instruction",
            "type": "string"
          },
          "temperature": {
            "default": 0,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "type": "number"
          }
        },
        "required": [
          "request_id",
          "product",
          "profile",
          "system_instruction",
          "input"
        ],
        "title": "TextGenerationRequest",
        "type": "object"
      },
      "TextGenerationResponse": {
        "additionalProperties": false,
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output": {
            "title": "Output",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageMeasurement"
          }
        },
        "required": [
          "request_id",
          "output",
          "provider",
          "model"
        ],
        "title": "TextGenerationResponse",
        "type": "object"
      },
      "UsageMeasurement": {
        "additionalProperties": false,
        "properties": {
          "audio_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Ms"
          },
          "input_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          }
        },
        "title": "UsageMeasurement",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Dali AI Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/ai/v1/audio/speech": {
      "post": {
        "operationId": "synthesize_speech_ai_v1_audio_speech_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dali-Caller",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dali-Caller"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechSynthesisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "audio/wav": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Synthesize Speech",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/v1/audio/transcriptions": {
      "post": {
        "operationId": "transcribe_audio_ai_v1_audio_transcriptions_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dali-Caller",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dali-Caller"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_transcribe_audio_ai_v1_audio_transcriptions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioTranscriptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transcribe Audio",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/v1/media/analyses": {
      "post": {
        "operationId": "analyze_media_ai_v1_media_analyses_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dali-Caller",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dali-Caller"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_media_ai_v1_media_analyses_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Media",
        "tags": [
          "AI"
        ]
      }
    },
    "/ai/v1/text/generations": {
      "post": {
        "operationId": "generate_text_ai_v1_text_generations_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Dali-Caller",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dali-Caller"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextGenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Text",
        "tags": [
          "AI"
        ]
      }
    },
    "/health/live": {
      "get": {
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Live Health Live Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Live",
        "tags": [
          "Health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ready Health Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready",
        "tags": [
          "Health"
        ]
      }
    }
  }
}